+++ RELEASES.md
@@ -41,6 +41,7 @@ Upcoming
 - Install a Servlet Filter to protect the Struts 1 Servlet from ClassLoader Manipulation attacks
 - `(end of v6.1-RC1)`
 - Improve IBIS console function "Show Adapter Statistics"
+- Add attribute simple to ExchangeMailListener to save memory